+++ changelog.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Perform new borrowing feature test.
 * Validation
     * Added maximal value for borrowing guarantee.
+* End borrowing view
+    * Added buttons blocking when no borrowing is selected.
 * Readme
     * Basic use guide.